pyproject.toml
[tool.poetry]
name = "hr-bot"
version = "0.1.0"
description = ""
authors = ["edfarrow <edvardfrrw@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
pydantic-settings = "^2.12.0"
google-generativeai = "^0.8.5"
bs4 = "^0.0.2"
aiogram = "^3.23.0"
fastapi = "^0.124.3"
redis = "^7.1.0"
httpx = "^0.28.1"
pypdf = "^6.4.1"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.38.0"
black = "^25.12.0"
flake8 = "^7.3.0"
coverage = "^7.13.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-mock = "^3.15.1"
pytest-cov = "^7.0.0"
coverage-badge = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
